filters pending conflicts
yet to sort by time created due to lack of sufficient data but now the modals are able to sieve out which data conflicts with the current booking req
